+++ cloudvolume/frontends/__init__.py
@@ -1,0 +1,2 @@
+from .precomputed import CloudVolumePrecomputed
+from .graphene import CloudVolumeGraphene
